MimeMagic: add this gem because remote content-type headers often lie
Manytimes a HEAD request to an arcgis server will return "text/html" when requesting a thumbnail. When you actually request the body, you end up with a png. So! We cannot trust the remote server to tell us what indeed they'll send us. MimeMagic allows us to inspect the filetype via the contents of the tempfile we generate, so we get the REAL mime, not the lie.